Deep filter for any association in defaultResolver
