+++ packages/core/package.json
@@ -61,7 +61,7 @@
     }
   },
   "dependencies": {
-    "graphql-lookahead": "^1.1.3"
+    "graphql-lookahead": "^1.2.0-beta.1"
   },
   "devDependencies": {
     "@types/node": "^20.17.8",
