+++ packages/plugin-sequelize/package.json
@@ -50,7 +50,7 @@
     }
   },
   "dependencies": {
-    "graphql-lookahead": "^1.1.3"
+    "graphql-lookahead": "^1.2.0-beta.1"
   },
   "devDependencies": {
     "@types/node": "^20.17.8",
